added tab bar, refactored views into distinct groups
