+++ 15puzzle/Views/Options/Options.swift
@@ -1,0 +1,18 @@
+//
+//  Options.swift
+//  15puzzle
+//
+//  Created by kai delay on 05/08/2023.
+//
+
+import SwiftUI
+
+struct Options: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    Options()
+}
